+++ tests/wiseagents/test_sequential_coordinator.py
@@ -2,9 +2,10 @@ import os
 import threading
 
 from wiseagents import WiseAgentMessage, WiseAgentRegistry
+from wiseagents.agents import SequentialCoordinatorWiseAgent
 from wiseagents.llm import OpenaiAPIWiseAgentLLM
 from wiseagents.transports import StompWiseAgentTransport
-from wiseagents.wise_agent_impl import LLMOnlyWiseAgent, PassThroughClientAgent, SequentialCoordinatorWiseAgent
+from wiseagents.wise_agent_impl import LLMOnlyWiseAgent, PassThroughClientAgent
 
 cond = threading.Condition()
